fix: rename exported shuvi-type-extensions files
