+++ packages/compiler/package.json
@@ -2,7 +2,7 @@
   "name": "@shuvi/compiler",
   "version": "1.0.0",
   "main": "./lib/index.js",
-  "types": "./lib/index.d.js",
+  "types": "./lib/index.d.ts",
   "files": [
     "lib"
   ],
