+++ packages/platform-shared/package.json
@@ -13,7 +13,7 @@
     "esm",
     "lib",
     "template",
-    "shuvi-type-extensions-node.d.js",
+    "shuvi-type-extensions-node.d.ts",
     "shuvi-type-extensions-node.js",
     "shuvi-type-extensions-runtime.d.ts"
   ],
@@ -64,7 +64,7 @@
     },
     "./template/*": "./template/*",
     "./shuvi-type-extensions-node": {
-      "types": "./shuvi-type-extensions-node.d.js",
+      "types": "./shuvi-type-extensions-node.d.ts",
       "default": "./shuvi-type-extensions-node.js"
     },
     "./shuvi-type-extensions-runtime": "./shuvi-type-extensions-runtime.d.ts",
