+++ packages/platform-shared/src/node/platform/shuvi-type-extensions-node.ts
@@ -1,0 +1,16 @@
+import { extendedHooks } from './plugins/main/hooks';
+import { IRuntimeConfig } from '../../shared';
+
+declare global {
+  namespace ShuviService {
+    interface CustomConfig {
+      publicRuntimeConfig?: IRuntimeConfig;
+      serverRuntimeConfig?: IRuntimeConfig;
+    }
+
+    interface CustomCorePluginHooks {
+      addEntryCode: typeof extendedHooks.addEntryCode;
+      modifyRuntimeConfig: typeof extendedHooks.modifyRuntimeConfig;
+    }
+  }
+}
