+++ packages/platform-shared/tsconfig.build.node.json
@@ -7,6 +7,5 @@
     "rootDir": "./src",
     "outDir": "lib"
   },
-  "files": ["./shuvi-type-extensions-node.d.ts"],
   "include": ["src/node", "src/shared"]
 }
